media album policy
- nisam nasla nacin kako da u policy proslijedim collection ili polje, ali skicnula sam mozda bi se moglo nesto skombinirati s gateom (s tim da u dokumentaciji pise da je policy za ozbiljniju autorizaciju, pa ne znam dal je u indexu i ovo s each okej)

- druga situacija je da sam bila dodala test za store metodu u proslom commitu za provjeriti da mi radi autorizacija radi (ne znam dal je taj test overkill i fale li mi mozda dodatni testovi - assert status 403 za ostale metode u kontroleru)

+++ app/Http/Controllers/Api/MediaAlbumController.php
@@ -18,6 +18,8 @@ class MediaAlbumController extends Controller
     {
         $mediaAlbums = MediaAlbum::whereIn('id', $request->validated('album_ids'))->get();
 
+        $mediaAlbums->each(fn($album) => $this->authorize('ownsAlbum', $album));
+
         return MediaAlbumsResource::collection($mediaAlbums);
     }
 
@@ -28,7 +30,7 @@ class MediaAlbumController extends Controller
             ['user_id' => auth()->id()]
         );
 
-        $this->authorize('store', $mediaAlbum);
+        $this->authorize('ownsAlbum', $mediaAlbum);
 
         $media = $request->safe()->collect('files')->map(fn($file) =>
         $mediaAlbum->addMedia($file)->toMediaCollection()
@@ -39,6 +41,8 @@ class MediaAlbumController extends Controller
 
     public function show(MediaAlbum $mediaAlbum)
     {
+        $this->authorize('ownsAlbum', $mediaAlbum);
+
         return MediaAlbumResource::collection(
             $mediaAlbum->media()->paginate()
         );
@@ -46,6 +50,8 @@ class MediaAlbumController extends Controller
 
     public function destroy(MediaAlbum $mediaAlbum)
     {
+        $this->authorize('ownsAlbum', $mediaAlbum);
+
         $mediaAlbum->media()->delete();
         $mediaAlbum->delete();
